third_party/blink/web_tests/platform/generic/svg/animations/svglength-animation-px-to-exs-expected.txt

SVG 1.1 dynamic animation tests

Test SVGLength animation from px to exs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 150
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 200
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
